Move GlyphBoundsAccumulator into TextBlock
